stock_exchange/requirements.txt
Django==2.0.2
django-webpack-loader==0.6.0
djangorestframework==3.7.7
pytz==2018.3
sqlparse==0.2.4
